More presice control of sleeps, now they are done in epoll_wait().

+++ engine.h
@@ -5,4 +5,4 @@
 
 struct state;
 bool transfer(size_t buffer_size, size_t block_size,
-              long sleep_us, struct state *const state);
+              int sleep_us, struct state *const state);
